flask_server/requirements.txt
alembic==1.7.5
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
bcrypt==3.2.0
cachelib==0.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
Deprecated==1.2.13
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.10.0
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
packaging==21.3
pycparser==2.21
pyparsing==3.0.6
python-dotenv==0.19.2
redis==4.1.0
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.29
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
WTForms==3.0.1
zipp==3.6.0
